Separated the CommandHandler
Divided the server's CommandHandler class into partial classes.
